requirements.txt
ccxt>=4.3.0
pandas>=2.2.0
python-dotenv>=1.0.0
